+++ src/setupTests.js
@@ -4,6 +4,9 @@
 // learn more: https://github.com/testing-library/jest-dom
 import "@testing-library/jest-dom";
 import { server } from "./__tests__/__mocks__/api/server";
+import setupStore from "./store/index";
+
+const store = setupStore({});
 
 // Establish API mocking before all tests.
 beforeAll(() => {