+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add MSW
 + Add util convert function
 + Add databases options function
++ Add Redux Tool Kit Query
 
 ## 1.4.0
 + Upgrade dependencies + refactor Semaphore and Docker